Fixes path expansion on Windows
A statement such as:

    let tmp = home(path.join('~/', '.tmp', '.nyc_output'));

On Windows will result in a path that looks like:

    '~\tmp\.nyc_output'

Using path.join() will change the path separator in the string, so the code fails to expand using the `USERPROFILE` directory.

The existing code only accepts a string where the first two characters are `~/`.  This change will search for both `~/` and `~\` for the home expansion.  It also changes the test runner to ava (prova didn't work for me on Windows).